+++ crates/installer/src/downloader/sources/mod.rs
@@ -1,8 +1,9 @@
 //! Download source implementations
 //!
 //! This module contains the actual implementation logic for different download sources.
-//! The data types are defined in `crate::parse_wabbajack::sources`.
+//! The data types are defined in `sources.rs`.
 
+pub mod sources;
 pub mod http;
 pub mod wabbajack_cdn;
 pub mod gamefile;
@@ -11,4 +12,5 @@ pub mod manual;
 pub mod archive;
 pub mod unknown;
 
-// Implementation modules (no re-exports needed since implementations are done via traits)
+// Re-export the types for cleaner imports
+pub use sources::*;
