+++ crates/installer/src/downloader/sources/unknown.rs
@@ -1,5 +1,5 @@
 use crate::downloader::core::{DownloadRequest, DownloadResult, ProgressCallback, Result, DownloadError};
-use crate::parse_wabbajack::sources::UnknownSource;
+use crate::downloader::sources::UnknownSource;
 
 impl UnknownSource {
     pub async fn download(&self, _request: &DownloadRequest, _progress_callback: Option<ProgressCallback>, _config: &crate::downloader::config::DownloadConfig) -> Result<DownloadResult> {
