+++ crates/installer/src/parse_wabbajack/mod.rs
@@ -7,11 +7,10 @@
 //! rather than converting to URL strings, as this provides better type safety,
 //! performance, and allows for richer data representation.
 
-pub mod sources;
 pub mod operations;
 pub mod parser;
 
 // Re-export main types
-pub use sources::{DownloadSource, HttpSource, NexusSource, GameFileSource, ManualSource, ArchiveSource};
+pub use crate::downloader::sources::{DownloadSource, HttpSource, NexusSource, GameFileSource, ManualSource, ArchiveSource};
 pub use operations::{ArchiveManifest, ManifestMetadata};
 pub use parser::{parse_modlist, ModlistParser};
